factor out encoded texture utilities into separate file
